+++ custom_components/senseme/const.py
@@ -6,3 +6,12 @@ DISCOVERY_UPDATE_RATE = 5
 
 # Periodic fan update rate in minutes
 UPDATE_RATE = 1
+
+# Custom config changes event
+EVENT_SENSEME_CONFIG_UPDATE = "senseme_config_updated"
+
+# Options
+CONF_ENABLE_WHOOSH = "enable_whoosh"
+CONF_ENABLE_WHOOSH_DEFAULT = True
+CONF_ENABLE_DIRECTION = "enable_direction"
+CONF_ENABLE_DIRECTION_DEFAULT = True